mebelniy/requirements.txt
aiogram>=3.2.0
python-dotenv==1.0.0
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
asyncpg==0.29.0

